Rename sheet_extent to sheet_position
