+++ lib/src/foundation/sheet_context.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:meta/meta.dart';
 
-import 'sheet_extent.dart';
+import 'sheet_position.dart';
 
 /// An interface that provides a set of dependencies required by [SheetExtent].
 @internal
