+++ lib/src/foundation/sheet_notification.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/widgets.dart';
 
 import 'sheet_drag.dart';
-import 'sheet_extent.dart';
 import 'sheet_physics.dart';
+import 'sheet_position.dart';
 import 'sheet_status.dart';
 
 /// A [Notification] that is dispatched when the sheet extent changes.
