+++ lib/src/navigation/navigation_routes.dart
@@ -4,10 +4,10 @@ import 'package:flutter/material.dart';
 import '../draggable/draggable_sheet_extent.dart';
 import '../draggable/draggable_sheet_extent_scope.dart';
 import '../draggable/sheet_draggable.dart';
-import '../foundation/sheet_extent.dart';
 import '../foundation/sheet_extent_scope.dart';
 import '../foundation/sheet_gesture_tamperer.dart';
 import '../foundation/sheet_physics.dart';
+import '../foundation/sheet_position.dart';
 import '../foundation/sheet_theme.dart';
 import '../scrollable/scrollable_sheet.dart';
 import '../scrollable/scrollable_sheet_extent.dart';
