+++ lib/src/navigation/navigation_sheet_extent.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';
 
 import '../foundation/sheet_drag.dart';
-import '../foundation/sheet_extent.dart';
+import '../foundation/sheet_position.dart';
 import '../internal/transition_observer.dart';
 import 'navigation_route.dart';
 import 'navigation_sheet_activity.dart';
