+++ lib/src/navigation/navigation_sheet_extent_scope.dart
@@ -1,9 +1,9 @@
 import 'package:meta/meta.dart';
 
 import '../foundation/sheet_context.dart';
-import '../foundation/sheet_extent.dart';
 import '../foundation/sheet_extent_scope.dart';
 import '../foundation/sheet_physics.dart';
+import '../foundation/sheet_position.dart';
 import 'navigation_sheet_extent.dart';
 
 @internal
